Pets part 1

+++ Intersect.Server/Migrations/PlayerContextModelSnapshot.cs
@@ -75,7 +75,7 @@ namespace Intersect.Server.Migrations
                     b.Property<Guid>("Id")
                         .ValueGeneratedOnAdd();
 
-                    b.Property<int>("Expires");
+                    b.Property<double>("Expires");
 
                     b.Property<Guid>("HDVId");
